\name{BSSprep-package}
\alias{BSSprep-package}
\docType{package}
\title{
Whitening Data as Preparation for Blind Source Separation
}

\description{
Whitening is the first step of almost all blind source separation (BSS) methods. A fast implementation of
whitening for BSS is implemented to serve as a lightweight dependency for packages providing BSS methods.}

\details{
\tabular{ll}{
Package: \tab BSSprep\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2021-03-25\cr
License: \tab GPL (>= 2)
}
This package contains the single function \code{\link{BSSprep}} for whitening multivariate data as a preprocessing step for blind source separation (BSS). The package is meant as a fast and lightweight dependency for packages providing BSS methods as whitening is almost always the first step.
}

\author{
Markus Matilainen, Klaus Nordhausen

Maintainer: Markus Matilainen <markus.matilainen@outlook.com> 
}

\keyword{package}
\keyword{multivariate}